AI-Tools/prompt_tester/requirements.txt
fastapi
uvicorn
requests
pydantic
streamlit
pandas
python-dotenv
tabulate